requirements.txt
streamlit
langchain
langchain-core
langchain_groq
requests
python-dotenv